Move server setup to a separate package
